backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
pandas==2.3.0
numpy
scikit-learn
python-dateutil
openpyxl==3.1.5
python-dotenv==1.0.1
werkzeug==3.1.3
gunicorn
requests